+++ src/clj/x/server_installer/api.clj
@@ -1,9 +1,30 @@
 
+;; -- Header ------------------------------------------------------------------
+;; ----------------------------------------------------------------------------
 
-- Szerver indulásakor megvizsgálja, hogy installálva van-e az X.
-- Ha nincs, akkor megkérdezi a konzol, hogy mehet-e az install.
-- Kiírja, hogy ne felejtsd el beállítani a project-details.edn fájlt.
-- Létrehozza a mongo-db kollekciókat.
-- Létrehozza a user-eket.
-- Létrehozza a project-details.edn fájlt (végre nem kell a mono-template-ben is up to date tartani)
-- Létrehozza a media mappákat.
+; Author: bithandshake
+; Created: 2021.10.18
+; Description:
+; Version: v0.2.2
+; Compatibility: x4.4.1
+
+
+
+;; -- Namespace ---------------------------------------------------------------
+;; ----------------------------------------------------------------------------
+
+(ns x.server-installer.api
+    (:require [x.server-installer.db-installer]
+              [x.server-installer.media-installer]
+              [x.server-installer.user-installer]
+              [x.server-installer.engine :as engine]))
+
+
+
+;; -- Redirects ---------------------------------------------------------------
+;; ----------------------------------------------------------------------------
+
+; x.server-installer.engine
+(def server-installed?     engine/server-installed?)
+(def get-installed-at      engine/get-installed-at)
+(def get-installed-version engine/get-installed-version)